+++ test/python/test_api.py
@@ -2393,9 +2393,10 @@ def test_terminate(tm):
     bitwuzla.assert_formula(b)
     assert bitwuzla.check_sat() == Result.UNKNOWN
 
-    # no terminator support in CryptoMiniSat and Kissat, but this will still
-    # terminate in the PP (as the terminator immediately terminates the
-    # execution on the first call to terminate)
+    # No terminator support in CryptoMiniSat, so configuring the terminator
+    # will already throw even though this would terminate in the PP (as the
+    # terminator immediately would terminate the execution on the first call to
+    # terminate).
     try:
         tt = TestTerminator()
         options.set(Option.BV_SOLVER, 'bitblast')
@@ -2403,10 +2404,14 @@ def test_terminate(tm):
         options.set(Option.REWRITE_LEVEL, 0)
         bitwuzla = Bitwuzla(tm, options)
         bitwuzla.configure_terminator(tt)
-        bitwuzla.assert_formula(b)
-        assert bitwuzla.check_sat() == Result.UNKNOWN
     except BitwuzlaException as e:
-        assert "CryptoMiniSat not compiled in" in str(e)
+        assert "CryptoMiniSat not compiled in" in str(e) or \
+               "terminator not supported in configured SAT solver" in str(e)
+
+    # No terminator support in Kissat, so configuring the terminator
+    # will already throw even though this would terminate in the PP (as the
+    # terminator immediately would terminate the execution on the first call to
+    # terminate).
     try:
         tt = TestTerminator()
         options.set(Option.BV_SOLVER, 'bitblast')
@@ -2414,10 +2419,9 @@ def test_terminate(tm):
         options.set(Option.REWRITE_LEVEL, 0)
         bitwuzla = Bitwuzla(tm, options)
         bitwuzla.configure_terminator(tt)
-        bitwuzla.assert_formula(b)
-        assert bitwuzla.check_sat() == Result.UNKNOWN
     except BitwuzlaException as e:
-        assert "Kissat not compiled in" in str(e)
+        assert "Kissat not compiled in" in str(e) or \
+               "terminator not supported in configured SAT solver" in str(e)
 
 def test_terminate_sat(tm):
     class TestTerminator: