+++ NEWS.md
@@ -2,6 +2,11 @@
 
 This file collects a summary of important and/or user-visible changes.
 
+- Configuring a terminator when a SAT solver that does not have terminator
+  support (CryptoMiniSat, Kissat) is configured will now raise an exception
+  in `Bitwuzla::configure_terminator()` rather than at `Bitwuzla::check_sat()`.
+  This is also handled correspondingly in the C and Python bindings.
+
 ## News for version 0.8.0
 
 - **Abstraction module** (option `--abstraction`) for abstracting bit-vector